english_text_analyzer/requirements.txt
# Core dependencies
langextract>=0.1.0
pydantic>=2.0.0
pyyaml>=6.0
typing-extensions>=4.0.0

# Optional dependencies for advanced features
# These will be used in later tasks
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0